+++ app/templates/components/value-input-fader.hbs
@@ -1,1 +1,1 @@
-{{yield}}
+{{~range-slider min=min max=max step=step value=value~}}
